+++ src/lib/logout/ProxyUserLogoutModal.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
 	import { onMount } from "svelte";
 	import Modal from "../Modal.svelte";
-	import axios from "axios";
 	import type { TrustedHeaderAuthLogoutDetailsResponse } from "@/types";
 	import Spinner from "../Spinner.svelte";
 	import { clearWatcharrData } from ".";
 	import { goto } from "$app/navigation";
+	import { req } from "../util/api";
 
 	interface Props {
 		onClose: () => void;
@@ -18,10 +18,10 @@
 
 	onMount(async () => {
 		try {
-			const r = await axios.get<TrustedHeaderAuthLogoutDetailsResponse>(
+			const r = await req.get<TrustedHeaderAuthLogoutDetailsResponse>(
 				"/auth/proxy_logout_details",
 			);
-			logoutUrl = r?.data?.logoutUrl;
+			logoutUrl = r?.logoutUrl;
 			if (!logoutUrl?.toLowerCase()?.startsWith("http")) {
 				// If no protocol in logoutUrl, set https
 				logoutUrl = `https://${logoutUrl}`;