+++ src/routes/(app)/+page.svelte
@@ -5,11 +5,11 @@
 	import Poster from "@/lib/poster/Poster.svelte";
 	import PosterList from "@/lib/poster/PosterList.svelte";
 	import Spinner from "@/lib/Spinner.svelte";
+	import { req } from "@/lib/util/api";
 	import infScroll from "@/lib/util/infScroll";
 	import paginatedLoader from "@/lib/util/paginatedLoader.svelte";
 	import { clearActiveFilters, store } from "@/store.svelte";
-	import type { Media } from "@/types";
-	import axios, { type GenericAbortSignal } from "axios";
+	import { type Media, type PaginationResponse } from "@/types";
 	import { onDestroy, untrack } from "svelte";
 
 	const scroll = infScroll({ callback: onScrollToBottom });
@@ -23,13 +23,13 @@
 		...store.sortAndFiltersForQueryParams,
 	});
 
-	async function load(signal: GenericAbortSignal) {
+	async function load(signal: AbortSignal) {
 		console.debug("load: loadParams:", nextLoadParams);
 		if (nextLoadParams.page === dataLoader.state.page) {
 			console.warn("load: Already on this page, not loading it again!");
 			return;
 		}
-		const r = await axios.get(`/watched`, {
+		const r = await req.get<PaginationResponse<Media, undefined>>(`/watched`, {
 			params: nextLoadParams,
 			signal,
 		});