+++ src/routes/(app)/arr_requests/+page.svelte
@@ -3,11 +3,10 @@
 	import Spinner from "@/lib/Spinner.svelte";
 	import RequestMovie from "@/lib/request/RequestMovie.svelte";
 	import RequestShow from "@/lib/request/RequestShow.svelte";
-	import { baseURL } from "@/lib/util/api";
+	import { baseURL, req } from "@/lib/util/api";
 	import { toRelativeDate } from "@/lib/util/helpers";
 	import { notify } from "@/lib/util/notify";
 	import { type ArrRequestResponse, type Media } from "@/types";
-	import axios from "axios";
 
 	let allRequests: ArrRequestResponse[] | undefined = $state();
 	let showBeingApproved: Media | undefined = $state();
@@ -16,8 +15,7 @@
 
 	async function getRequests() {
 		try {
-			allRequests = (await axios.get<ArrRequestResponse[]>(`/arr/request/`))
-				.data;
+			allRequests = await req.get<ArrRequestResponse[]>(`/arr/request/`);
 			if (allRequests?.length > 0) {
 				allRequests = allRequests?.sort((a, b) => {
 					if (b.status === "PENDING") return 1;
@@ -33,7 +31,7 @@
 
 	async function deny(r: ArrRequestResponse) {
 		try {
-			await axios.post(`/arr/request/deny/${r.id}`);
+			await req.post(`/arr/request/deny/${r.id}`);
 			getRequests();
 		} catch (err) {
 			console.error("Failed to deny request!", err);
@@ -44,13 +42,13 @@
 	async function approve(r: ArrRequestResponse) {
 		console.debug("Approving request:", r);
 		if (r.content.type === "tv") {
-			showBeingApproved = (
-				await axios.get<Media>(`/content/tv/${r.content.tmdbId}`)
-			).data;
+			showBeingApproved = await req.get<Media>(
+				`/content/tv/${r.content.tmdbId}`,
+			);
 		} else if (r.content.type === "movie") {
-			movieBeingApproved = (
-				await axios.get<Media>(`/content/movie/${r.content.tmdbId}`)
-			).data;
+			movieBeingApproved = await req.get<Media>(
+				`/content/movie/${r.content.tmdbId}`,
+			);
 		} else {
 			notify({
 				type: "error",